requirements.txt
python-dotenv
google.generativeai
python-multipart
pdfminer.six
python-docx
pandas
nltks